mockscull/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "numscull"
version = "0.2.4"
description = "Numscull Python client — collaborative code-annotation with encrypted communication"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "jsonschema>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
